testdriven-app/services/leagues/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.8.6
SQLAlchemy-serializer
